+++ resources/js/Pages/components/centralPages/dashboard/subscription/AnaliyticsContent.vue
@@ -1,0 +1,3 @@
+<template>
+    <h1>Analystic Content</h1>
+</template>
