+++ resources/js/Pages/components/centralPages/dashboard/subscription/FeaturesContent.vue
@@ -1,0 +1,5 @@
+<template>
+    <div>
+        <h1>Features Content</h1>
+    </div>
+</template>
