+++ resources/js/Pages/views/dashboard/central_page/Subscription_page.vue
@@ -1,79 +1,49 @@
 <script setup>
-import { onMounted, reactive, ref } from 'vue';
+import { onMounted, defineAsyncComponent, ref, watch } from 'vue';
 import DashboardLayout from '@/layouts/DashboardLayout.vue';
 import { Head } from '@inertiajs/vue3';
 
 import { useNavMainPlatform } from '@/stores/navigation_menu_item';
 import { storeToRefs } from 'pinia';
 
-import Card from '@/elements/card/DefaultCard.vue';
-import DataTable from 'primevue/datatable';
-import Column from 'primevue/column';
-import Badge from 'primevue/badge';
-import Dropdown from 'primevue/dropdown';
-
-const dataSubscriptions = ref([
-    {
-        id: 1,
-        plan: "Silver plan",
-        features: [
-            "100 Member",
-            "5 Plan subscription"
-        ],
-        price: 50000,
-        typePlan: "Monthly"
-    },
-    {
-        id: 2,
-        plan: "Gold plan",
-        features: [
-            "100 Member",
-            "5 Plan subscription",
-            "PT Management"
-        ],
-        price: 1500000,
-        typePlan: "Monthly"
-    },
-    {
-        id: 3,
-        plan: "Platinum plan",
-        features: [
-            "100 Member",
-            "5 Plan subscription",
-            "PT Management",
-            "Open Class Mentoring"
-        ],
-        price: 2000000,
-        typePlan: "Yearly"
-    },
-    {
-        id: 4,
-        plan: "Ultimate plan",
-        features: [
-            "500 Member",
-            "5 Plan subscription",
-            "PT Management",
-            "Open Class Mentoring",
-            "Gym Stuff Management"
-        ],
-        price: 5000000,
-        typePlan: "Yearly"
-    },
-])
-
-const subscriptionActions = reactive(["Delete", "Edit", "Disabled"])
+import TabView from 'primevue/tabview';
+import TabPanel from 'primevue/tabpanel'
+import Button from 'primevue/button'
 
 const getNavMainPlatform = useNavMainPlatform()
 const { updateMenusItemActive } = getNavMainPlatform
 const { navigationMenuItem } = storeToRefs(getNavMainPlatform)
 
+const PlanTable = defineAsyncComponent({
+    loader: () => import('@dashboardCentralPageComponents/subscription/PlanTable.vue'),
+})
+
+const AnaliyticsPage = defineAsyncComponent({
+    loader: () => import('@dashboardCentralPageComponents/subscription/AnaliyticsContent.vue')
+})
+
+const FeaturesPage = defineAsyncComponent({
+    loader: () => import('@dashboardCentralPageComponents/subscription/FeaturesContent.vue')
+})
+
+const tabsContent = ref([
+    { title: "Analiytics", content: AnaliyticsPage, isActive: true },
+    { title: "Plans Table", content: PlanTable, isActive: false },
+    { title: "Features Table", content: FeaturesPage, isActive: false }
+])
+
+const contentTabActive = ref(0)
+
 onMounted(() => {
     updateMenusItemActive(route(route().current()))
 })
 
-const formatCurrency = (price) => {
-    return price.toLocaleString('id-ID', { style: "currency", currency: "IDR" })
-}
+watch(contentTabActive, (stateNow, prevState) => {
+    if (prevState) {
+        tabsContent.value[prevState].isActive = false
+    }
+    tabsContent.value[stateNow].isActive = true
+})
 
 </script>
 
@@ -84,65 +54,61 @@ const formatCurrency = (price) => {
     <DashboardLayout :menu-items="navigationMenuItem" titleNav="Subscription Management">
         <template #main_content>
 
-            <section class="table_subscriptions">
+            <TabView pt:nav="flex flex-1 bg-transparent list-none p-0 m-0 border-b-2 dark:border-primary-700" :pt="{
+                tabpanel: {
+                    headerAction: ({ parent, context }) => ({
+                        class: ['bg-transparent', 'relative',
+                            // Font
+                            'font-bold',
+                            // Flexbox and Alignment
+                            'flex items-center',
+                            // Spacing
+                            'p-5',
+                            '-mb-[2px]',
+                            // Shape
+                            'border-b-2',
+                            'rounded-t-md',
+                            {
+                                'border-b-2 border-transparent': parent.state.d_activeIndex !== context.index,
+                                'bg-transparent font-normal': parent.state.d_activeIndex !== context.index,
+                                'text-surface-700 dark:text-surface-0/80': parent.state.d_activeIndex !== context.index,
+                                'bg-transparent': parent.state.d_activeIndex === context.index,
+                                'border-surface-400 dark:border-surface-400': parent.state.d_activeIndex === context.index,
+                                'text-white dark:text-white': parent.state.d_activeIndex === context.index
+                            },
+                            {
+                                'hover:text-white': parent.state.d_activeIndex !== context.index,
+                                'hover:border-surface-400 dark:hover:border-surface-400': parent.state.d_activeIndex !== context.index
+                            },
+                            'focus-visible:outline-none focus-visible:outline-offset-0 focus-visible:ring focus-visible:ring-inset',
+                            'transition-all duration-200',
+                            'cursor-pointer select-none text-decoration-none',
+                            'user-select-none',
+                            'overflow-hidden'
+                        ]
+                    }),
+                    content: {
+                        class: ['bg-transparent py-5 border-0']
+                    }
+                }
+            }" v-model:activeIndex="contentTabActive">
 
-                <Card>
-                    <template #card__title>Tables</template>
-                    <template #card__content>
-                        <DataTable :value="dataSubscriptions" :pt="{
-        bodyrow: 'bg-transparent last:border-none border-b border-primary-700',
-        column: {
-            headercell: 'py-6 px-4 border-b border-primary-600',
-            headercontent: 'text-left font-[300] tracking-wide text-white/50',
-            bodycell: 'p-5',
-        }
-    }">
-                            <Column header="No">
-                                <template #body="slotProps">
-                                    <div>{{ slotProps.data.id }}</div>
-                                </template>
-                            </Column>
-                            <Column header="Plans">
-                                <template #body="slotProps">
-                                    <div>{{ slotProps.data.plan }}</div>
-                                </template>
-                            </Column>
-                            <Column header="Prices">
-                                <template #body="slotProps">
-                                    <div>{{ formatCurrency(slotProps.data.price) }}</div>
-                                </template>
-                            </Column>
-                            <Column header="Features">
-                                <template #body="slotProps">
-                                    <ul>
-                                        <li v-for="item in slotProps.data.features" class="mb-3">{{ item }}</li>
-                                    </ul>
-                                </template>
-                            </Column>
-                            <Column header="Types">
-                                <template #body="slotProps">
-                                    <Badge :value="slotProps.data.typePlan"
-                                        :severity="slotProps.data.typePlan === 'Monthly' ? `success` : `info`">
-                                    </Badge>
-                                </template>
-                            </Column>
-                            <Column header="Actions">
-                                <template #body="slotProps">
-                                    <Dropdown placeholder="Action" :options="subscriptionActions">
-                                        <template #value="slotProps">
-                                            <div v-if="slotProps.value">
-                                                {{ slotProps.value }}
-                                            </div>
-                                            <div v-else>{{ slotProps.placeholder }}</div>
-                                        </template>
-                                    </Dropdown>
-                                </template>
-                            </Column>
-                        </DataTable>
-                    </template>
-                </Card>
+                <TabPanel v-for="(tab) in tabsContent" :key="tab.title" :header="tab.title">
+                    <Suspense v-if="tab.isActive">
+                        <template #default>
+                            <component :is="tab.content"></component>
+                        </template>
 
-            </section>
+                        <template #fallback>
+                            <div>
+                                loading...
+                            </div>
+                        </template>
+                    </Suspense>
+                </TabPanel>
+
+            </TabView>
+
         </template>
     </DashboardLayout>
 </template>
